@@ -2,8 +2,8 @@ package lib.exoplayer;
 
 import android.content.Context;
 import android.graphics.Color;
-import android.support.annotation.NonNull;
-import android.support.v4.graphics.ColorUtils;
+import androidx.annotation.NonNull;
+import androidx.core.graphics.ColorUtils;
 import android.text.SpannableStringBuilder;
 import android.text.Spanned;
 import android.util.Log;
