@@ -3,9 +3,9 @@ package lib.exoplayer;
 import android.animation.TimeInterpolator;
 import android.animation.ValueAnimator;
 import android.os.Build;
-import android.support.annotation.FloatRange;
-import android.support.annotation.IntRange;
-import android.support.annotation.NonNull;
+import androidx.annotation.FloatRange;
+import androidx.annotation.IntRange;
+import androidx.annotation.NonNull;
 import android.text.TextPaint;
 import android.text.style.SuperscriptSpan;
 import android.view.View;
